Encapsulate sign in process for DRMs
